requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.20
openai==1.69.0
chromadb==0.4.24
httpx==0.27.0
python-dotenv==1.0.1
pydantic>=2,<3
protobuf==4.25.3
onnxruntime==1.18.0
langserve==0.3.1
python-multipart==0.0.9 
sse-starlette==1.6.1
